+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -154,6 +154,16 @@ const VirtualListBaseFactory = (type) => {
 			pageScroll: PropTypes.bool,
 
 			/**
+			 * The role for the list.
+			 * If unset, it defaults to the `list` of `role`
+			 *
+			 * @type {String}
+			 * @default 'list'
+			 * @publie
+			 */
+			role: PropTypes.string,
+
+			/**
 			 * `true` if rtl, `false` if ltr.
 			 * Normally, [Scrollable]{@link ui/Scrollable.Scrollable} should set this value.
 			 *
@@ -796,7 +806,7 @@ const VirtualListBaseFactory = (type) => {
 
 		render () {
 			const
-				{itemRenderer, itemsRenderer, ...rest} = this.props,
+				{itemRenderer, itemsRenderer, role, ...rest} = this.props,
 				needsScrollingPlaceholder = this.isNeededScrollingPlaceholder();
 
 			delete rest.initUiChildRef;
@@ -812,7 +822,7 @@ const VirtualListBaseFactory = (type) => {
 					getComponentProps={this.getComponentProps}
 					itemRenderer={({index, ...itemRest}) => ( // eslint-disable-line react/jsx-no-bind
 						itemRenderer({
-							... itemRest,
+							...itemRest,
 							[dataIndexAttribute]: index,
 							index
 						})
@@ -824,7 +834,8 @@ const VirtualListBaseFactory = (type) => {
 						return itemsRenderer({
 							...props,
 							handlePlaceholderFocus: this.handlePlaceholderFocus,
-							needsScrollingPlaceholder
+							needsScrollingPlaceholder,
+							role
 						});
 					}}
 				/>
@@ -930,13 +941,14 @@ const listItemsRenderer = (props) => {
 		handlePlaceholderFocus,
 		itemContainerRef: initUiItemContainerRef,
 		needsScrollingPlaceholder,
-		primary
+		primary,
+		role
 	} = props;
 
 	return (
 		<React.Fragment>
 			{cc.length ? (
-				<div ref={initUiItemContainerRef} role="list">{cc}</div>
+				<div ref={initUiItemContainerRef} role={role}>{cc}</div>
 			) : null}
 			{primary ? null : (
 				<SpotlightPlaceholder
@@ -956,22 +968,21 @@ const listItemsRenderer = (props) => {
 };
 /* eslint-enable enact/prop-types */
 
-const ScrollableVirtualList = (props) => { // eslint-disable-line react/jsx-no-bind
-	const {focusableScrollbar} = props;
-
+const ScrollableVirtualList = ({role, ...rest}) => { // eslint-disable-line react/jsx-no-bind
 	warning(
-		!props.itemSizes || !props.cbScrollTo,
+		!rest.itemSizes || !rest.cbScrollTo,
 		'VirtualList with `minSize` in `itemSize` prop does not support `cbScrollTo` prop'
 	);
 
 	return (
 		<Scrollable
-			{...props}
+			{...rest}
 			childRenderer={(childProps) => ( // eslint-disable-line react/jsx-no-bind
 				<VirtualListBase
 					{...childProps}
-					focusableScrollbar={focusableScrollbar}
+					focusableScrollbar={rest.focusableScrollbar}
 					itemsRenderer={listItemsRenderer}
+					role={role}
 				/>
 			)}
 		/>
@@ -983,31 +994,32 @@ ScrollableVirtualList.propTypes = /** @lends moonstone/VirtualList.VirtualListBa
 	direction: PropTypes.oneOf(['horizontal', 'vertical']),
 	focusableScrollbar: PropTypes.bool,
 	itemSizes: PropTypes.array,
-	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic'])
+	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+	role: PropTypes.string
 };
 
 ScrollableVirtualList.defaultProps = {
 	direction: 'vertical',
 	focusableScrollbar: false,
-	preventBubblingOnKeyDown: 'programmatic'
+	preventBubblingOnKeyDown: 'programmatic',
+	role: 'list'
 };
 
-const ScrollableVirtualListNative = (props) => {
-	const {focusableScrollbar} = props;
-
+const ScrollableVirtualListNative = ({role, ...rest}) => {
 	warning(
-		!props.itemSizes || !props.cbScrollTo,
+		!rest.itemSizes || !rest.cbScrollTo,
 		'VirtualList with `minSize` in `itemSize` prop does not support `cbScrollTo` prop'
 	);
 
 	return (
 		<ScrollableNative
-			{...props}
+			{...rest}
 			childRenderer={(childProps) => ( // eslint-disable-line react/jsx-no-bind
 				<VirtualListBaseNative
 					{...childProps}
-					focusableScrollbar={focusableScrollbar}
+					focusableScrollbar={rest.focusableScrollbar}
 					itemsRenderer={listItemsRenderer}
+					role={role}
 				/>
 			)}
 		/>
@@ -1019,13 +1031,15 @@ ScrollableVirtualListNative.propTypes = /** @lends moonstone/VirtualList.Virtual
 	direction: PropTypes.oneOf(['horizontal', 'vertical']),
 	focusableScrollbar: PropTypes.bool,
 	itemSizes: PropTypes.array,
-	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic'])
+	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+	role: PropTypes.string
 };
 
 ScrollableVirtualListNative.defaultProps = {
 	direction: 'vertical',
 	focusableScrollbar: false,
-	preventBubblingOnKeyDown: 'programmatic'
+	preventBubblingOnKeyDown: 'programmatic',
+	role: 'list'
 };
 
 export default VirtualListBase;