+++ packages/ui/CHANGELOG.md
@@ -12,6 +12,7 @@ No significant changes.
 
 - `ui/Scroller` TypeScript signatures
 - `ui/VirtualList.VirtualGridList` and `ui/VirtualList.VirtualList` to apply `will-change` CSS property to the proper node
+- `moonstone/VirtualList.VirtualGridList` and `moonstone/VirtualList.VirtualList` ability to change of the role
 
 ## [3.0.0-rc.4] - 2019-08-22